% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_auxiliary.r
\name{apply_missingness}
\alias{apply_missingness}
\title{Apply missingness to data}
\usage{
apply_missingness(x, prop)
}
\arguments{
\item{x}{Vector or column to apply missingness to}

\item{prop}{Proportion of values to set as missing}
}
\value{
Vector with NA values
}
\description{
Introduces missing values (NA) at random
}
