% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bace_option_setter_getter.r
\name{bace_option_defaults}
\alias{bace_option_defaults}
\title{default option values}
\usage{
bace_option_defaults()
}
\value{
named list of default option values
}
\description{
Internal function returning default option values. Currently:
\itemize{
\item verbose: TRUE
\item digits: 3L
\item gelman: 2 (pseudo-Gelman prior scaling)
}
}
