% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_auxiliary.r
\name{beta_generator}
\alias{beta_generator}
\title{Generate beta coefficients from beta matrix row}
\usage{
beta_generator(beta_row, ns, default_beta = 0.5)
}
\arguments{
\item{beta_row}{Row of beta matrix for the focal variable}

\item{ns}{Vector of parameter counts per predictor}

\item{default_beta}{Default beta value if using placeholders}
}
\value{
Vector of beta coefficients
}
\description{
Extracts and expands beta coefficients for a variable. Uses the
beta matrix and response betas to generate the full set of coefficients
needed for predictors with multiple parameters. For cases of multiple categories,
the current versions is just spreading the effect evenly across levels.
}
