% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_functions.R
\name{check_type}
\alias{check_type}
\title{check_type}
\usage{
check_type(x)
}
\arguments{
\item{x}{A vector of data}
}
\value{
Returns a character string specifying the class of the variable.
}
\description{
Function takes a variable in a dataframe and checks/ classifies what tope
}
