% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_auxiliary.r
\name{design_size}
\alias{design_size}
\title{Calculate design size and simulation order}
\usage{
design_size(predictor_types, beta_matrix)
}
\arguments{
\item{predictor_types}{Character vector of predictor types}

\item{beta_matrix}{Matrix indicating dependencies between predictors}
}
\value{
List with ns (number of parameters per predictor) and sim_order
}
\description{
Determines the number of parameters needed per predictor
and the order in which predictors should be simulated based on dependencies.
To avoid numerical issues - the first predictor should be continuous gaussian.
}
