% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{.build_formula}
\alias{.build_formula}
\title{.build_formula}
\usage{
.build_formula(x)
}
\arguments{
\item{x}{A character string specifying the formula used in the model. This should be of the form: y ~ x.}
}
\value{
A formula object
}
\description{
Function takes a string specifying a formula and converts this to a formula object to be used in the models.
}
\examples{
\dontrun{
.build_formula("y ~ x1 + x2")
}
}
