% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{.check_mcmc_diagnostics}
\alias{.check_mcmc_diagnostics}
\title{.check_mcmc_diagnostics}
\usage{
.check_mcmc_diagnostics(bace_output)
}
\arguments{
\item{bace_output}{Output from bace_imp function containing models_last_run}
}
\value{
A list containing diagnostic statistics and a summary data frame
}
\description{
Function checks MCMC convergence and mixing for all parameters in MCMCglmm models
}
\details{
This function performs the following diagnostics:
\itemize{
\item Effective Sample Size (ESS) for all parameters
\item Geweke convergence diagnostic
\item Autocorrelation at lag 1
\item Summary statistics for fixed and random effects
}
}
