% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_functions.R
\name{.data_prep}
\alias{.data_prep}
\title{data_prep}
\usage{
.data_prep(formula, data, types, ran_cluster)
}
\arguments{
\item{formula}{A formula specifying the response and predictor variables.}

\item{data}{A data frame containing the dataset to be prepared.}

\item{types}{A list specifying the type of each variable in the dataset.}

\item{ran_cluster}{A string specifying the random effect cluster variable.}
}
\value{
A list containing the prepared data frame and attributes for continuous variables.
}
\description{
Function prepares the data for imputation by handling missing values and standardizing continuous variables.
}
\examples{
\dontrun{
data <- data.frame(y = c(1,2,3,NA,5), x1 = factor(c("A","B","A","B","A")), x2 = c(10,20,30,NA,50))
formula <- as.formula("y ~ x1 + x2")
types <- list(y = "gaussian", x1 = "categorical", x2 = "gaussian")
data_prep(formula, data, types)
}	
}
