% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_diagnostics.R
\name{.energy_distance}
\alias{.energy_distance}
\title{.energy_distance}
\usage{
.energy_distance(x1, x2)
}
\arguments{
\item{x1}{Matrix or data frame (first sample)}

\item{x2}{Matrix or data frame (second sample)}
}
\value{
Numeric energy distance
}
\description{
Calculate energy distance between two multivariate samples
}
