% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_functions.R
\name{.extract_gaussian_attrs}
\alias{.extract_gaussian_attrs}
\title{extract_gaussian_attrs}
\usage{
.extract_gaussian_attrs(data, types)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{types}{A list specifying the type of each variable in the dataset.}
}
\value{
A list containing the mean and standard deviation of each gaussian variable.
}
\description{
Function extracts the mean and standard deviation of gaussian variables in a dataframe for later use in back-transformation.
}
\examples{
\dontrun{
data <- data.frame(y = c(1,2,3,NA,5), x1 = factor(c("A","B","A","B","A")), x2 = c(10,20,30,NA,50))
types <- list(y = "gaussian", x1 = "categorical", x2 = "gaussian")
extract_gaussian_attrs(data, types)
}
}
