% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_imputations.R
\name{.extract_imputed_datasets}
\alias{.extract_imputed_datasets}
\title{Extract the list of completed datasets from a BACE object}
\usage{
.extract_imputed_datasets(object)
}
\description{
Extract the list of completed datasets from a BACE object
}
\keyword{internal}
