% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_functions.R
\name{.get_levels}
\alias{.get_levels}
\title{.get_levels}
\usage{
.get_levels(x, store_levels, max_levels_store)
}
\arguments{
\item{x}{A variable}

\item{store_levels}{A logical indicating whether to store levels}

\item{max_levels_store}{An integer specifying the maximum number of levels to store}
}
\value{
A vector of levels or NULL if not applicable
}
\description{
Helper function to get levels of a variable if applicable
}
