% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_functions.R
\name{.get_type}
\alias{.get_type}
\title{.get_type}
\usage{
.get_type(x, var, data)
}
\arguments{
\item{x}{A summary of a variables in a dataframe}

\item{var}{The name of the variable}

\item{data}{The dataframe containing the variables}
}
\value{
Returns a character string specifying the class of the variable.
}
\description{
Function takes a variable in a dataframe and checks / classifies what type of variable it is (binary, continuous, count, categorical, ordered_categorical) so that it can be modeled appropriately.
}
\keyword{internal}
