% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{.impute_levels}
\alias{.impute_levels}
\title{.impute_levels}
\usage{
.impute_levels(pred_prob, levels_var, sample = FALSE)
}
\arguments{
\item{pred_prob}{A data frame of predicted probabilities for each category}

\item{levels_var}{A character vector specifying the names of the levels/categories}

\item{sample}{A logical indicating whether to sample from the distribution or take the maximum probability}
}
\value{
A vector of predicted levels for each observation
}
\description{
Function samples predicted levels for categorical or ordinal variables based on predicted probabilities
}
