% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{.list_of_G}
\alias{.list_of_G}
\title{.list_of_G}
\usage{
.list_of_G(n_rand, nu = NULL, par_expand = TRUE, diag = 1, alpha_V = 10000)
}
\arguments{
\item{n_rand}{An integer specifying the number of random effects in the model.}

\item{nu}{A numeric specifying the nu parameter for the prior.}

\item{par_expand}{A logical indicating whether to use parameter expansion.}

\item{diag}{An integer specifying the dimension of the variance-covariance matrix.}

\item{alpha_V}{Scale of the parameter-expanded working-prior variance on the
random-effect standard deviation (alpha.V = alpha_V * I). Default 1e4
(identity-link scales); log-link callers (poisson) pass 1.}
}
\value{
A list of G priors for the MCMCglmm model.
}
\description{
Function creates a list of G priors for the MCMCglmm model
}
