% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{.make_prior}
\alias{.make_prior}
\title{.make_prior}
\usage{
.make_prior(n_rand, type, nu = NULL, n_levels = NULL, par_expand = FALSE)
}
\arguments{
\item{n_rand}{An integer specifying the number of random effects in the model.}

\item{type}{A string that specifies the type of model to fit.}

\item{nu}{A numeric specifying the nu parameter for the prior.}

\item{n_levels}{An integer specifying the number of levels for categorical or ordinal data.}

\item{par_expand}{A logical indicating whether to use parameter expansion.}
}
\value{
A list of priors for the MCMCglmm model.
}
\description{
Function creates the prior for the MCMCglmm model
}
