% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_mi.R
\name{.mcmcglmm_fixef_idx}
\alias{.mcmcglmm_fixef_idx}
\title{Column indices of the fixed effects in an MCMCglmm Sol matrix}
\usage{
.mcmcglmm_fixef_idx(model)
}
\description{
Column indices of the fixed effects in an MCMCglmm Sol matrix
}
\keyword{internal}
