% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{.model_fit}
\alias{.model_fit}
\title{.model_fit}
\usage{
.model_fit(
  data,
  tree,
  fixformula,
  randformula,
  type,
  prior,
  nitt = 6000,
  thin = 5,
  burnin = 1000
)
}
\arguments{
\item{data}{A dataframe containing missing data.}

\item{tree}{A phylogenetic tree of class 'phylo' from the ape package.}

\item{fixformula}{A string that specifies the fixed effect structure in the model.}

\item{randformula}{A formula or list of formulas that specifies the random effect structure in the model. If a list, should have elements named 'phylo' and 'species'.}

\item{type}{A string that specifies the type of model to fit. Options are "normal", "binary", "count", "categorical", "ordinal" and the appropriate family will be used in MCMCglmm.}

\item{prior}{A list specifying the prior distributions for the MCMCglmm model.}

\item{nitt}{An integer specifying the number of iterations to run the MCMC algorithm for this specific model. Default is 6000. Note: when called from bace_imp, this can be model-specific if a list was provided.}

\item{thin}{An integer specifying the thinning rate for the MCMC algorithm for this specific model. Default is 5. Note: when called from bace_imp, this can be model-specific if a list was provided.}

\item{burnin}{An integer specifying the number of iterations to discard as burnin for this specific model. Default is 1000. Note: when called from bace_imp, this can be model-specific if a list was provided.}
}
\value{
A list of draws from the posterior distribution of the model parameters.
}
\description{
Function takes incomplete data and a formula string and fits a model using MCMCglmm
}
