% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{.pred_cat}
\alias{.pred_cat}
\title{.pred_cat}
\usage{
.pred_cat(model, baseline_name = "Baseline")
}
\arguments{
\item{model}{A MCMCglmm model object}

\item{baseline_name}{A string specifying the name of the baseline category}
}
\value{
A data frame of predicted probabilities for each category
}
\description{
Function calculates predicted probabilities for each category from a categorical MCMCglmm model
}
