% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{.pred_threshold}
\alias{.pred_threshold}
\title{.pred_threshold}
\usage{
.pred_threshold(model, level_names = NULL)
}
\arguments{
\item{model}{A MCMCglmm model object}

\item{level_names}{A character vector specifying the names of the levels/categories}
}
\value{
A data frame of predicted probabilities for each category
}
\description{
Function calculates predicted probabilities for each category from a threshold MCMCglmm model
}
