% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{.replace_symbols}
\alias{.replace_symbols}
\title{.replace_symbols}
\usage{
.replace_symbols(expr, map)
}
\arguments{
\item{expr}{An expression}

\item{map}{A named list specifying the replacements to be made}
}
\value{
An expression with the symbol names replaced
}
\description{
Recursively replace symbol names in an expression
}
