% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bace_imp.R
\name{.standardize_mcmc_params}
\alias{.standardize_mcmc_params}
\title{.standardize_mcmc_params}
\usage{
.standardize_mcmc_params(param, n_models, param_name)
}
\arguments{
\item{param}{The parameter value (either a single value or a list)}

\item{n_models}{The number of models/formulas}

\item{param_name}{The name of the parameter (for error messages)}
}
\value{
A list of length n_models with the parameter values
}
\description{
Helper function to standardize MCMC parameters (nitt, thin, burnin) to lists
}
