% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{.swap_symbols}
\alias{.swap_symbols}
\title{.swap_symbols}
\usage{
.swap_symbols(expr, a, b, tmp = ".TMP_SWAP_SYMBOL")
}
\arguments{
\item{expr}{An expression}

\item{a}{The first symbol name to swap}

\item{b}{The second symbol name to swap}

\item{tmp}{A temporary symbol name used during the swap}
}
\value{
An expression with the two symbol names swapped
}
\description{
Swap two symbol names in an expression
}
