% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_diagnostics.R
\name{.wasserstein_1d}
\alias{.wasserstein_1d}
\title{.wasserstein_1d}
\usage{
.wasserstein_1d(x, y)
}
\arguments{
\item{x}{First sample}

\item{y}{Second sample}
}
\value{
Numeric Wasserstein distance
}
\description{
Calculate 1-dimensional Wasserstein distance
}
