% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_auxiliary.r
\name{generate_default_beta_matrix}
\alias{generate_default_beta_matrix}
\title{Generate default beta matrix}
\usage{
generate_default_beta_matrix(
  n_predictors,
  sparsity = 0.7,
  beta_range = c(-0.5, 0.5)
)
}
\arguments{
\item{n_predictors}{Number of predictors}

\item{sparsity}{Proportion of zero entries (default 0.7)}

\item{beta_range}{Range for non-zero beta values}
}
\value{
Square matrix of beta coefficients
}
\description{
Creates a default beta matrix with dependency structure set
through the sparsity parameter. it produces a single beta coefficient per
predictor - in cases multiple values are needed the other function
beta_generator() will expand them accordingly.
}
