% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_auxiliary.r
\name{generate_ix_betas}
\alias{generate_ix_betas}
\title{Generate interaction coefficients}
\usage{
generate_ix_betas(interactions, beta_ix = NULL, beta_range = c(-0.3, 0.3))
}
\arguments{
\item{interactions}{List of interaction pairs (from parse_ix_matrix)}

\item{beta_ix}{User-specified interaction coefficients (named list or NULL)}

\item{beta_range}{Range for random coefficient generation}
}
\value{
Named list of interaction coefficients
}
\description{
Creates coefficients for interaction terms. Can be specified
directly or generated randomly.
}
