% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_functions.R
\name{get_class}
\alias{get_class}
\title{get_class}
\usage{
get_class(x)
}
\arguments{
\item{x}{A dataframe containing missing data.}
}
\value{
A list identifying the class of all columns in the dataframe.
}
\description{
Function identifies the class of all variables in the dataframe
}
\examples{
{
data(iris)
get_class(iris)}
}
