% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{get_imputed}
\alias{get_imputed}
\title{get_imputed}
\usage{
get_imputed(model, type)
}
\arguments{
\item{model}{An integer specifying the number of random effects in the model.}

\item{type}{A string that specifies the type of model to fit.}
}
\value{
imputated data based the MCMCglmm model.
}
\description{
Function extracts the prior for the MCMCglmm model
}
