% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_functions.R
\name{get_type}
\alias{get_type}
\title{get_type}
\usage{
get_type(data)
}
\arguments{
\item{fixed_form}{The fixed effect formula}

\item{cluster_name}{The name of the cluster variable}

\item{random_form}{The random effect formula}
}
\value{
A list of variable names and their corresponding class.
}
\description{
Function takes a dataframe and identifies the class of variables (type)
}
\examples{
\dontrun{
data <- data.frame(y = c(1,2,3,NA,5), x1 = factor(c("A","B","A","B","A")), x2 = c(10,20,30,40,50))
get_type(data)
}
}
