% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_functions.R
\name{get_variables}
\alias{get_variables}
\title{get_variables}
\usage{
get_variables(x)
}
\arguments{
\item{x}{A string}
}
\value{
A vector of variable names. This vector is used to subset out the 'fixed effects' and 'random effect' columns within a given dataframe.
}
\description{
Function takes a formula string and identifies the variables in the formula that relate to the data
}
\examples{
{
# All should return the same three variables: y, x1, x2
form <- "y ~ x1 + x2"
get_variables(form)
form <- "y ~ x1 * x2"
get_variables(form)
form <- "y ~ x1 + x2 + x1 : x2"
get_variables(form)
form <- "y ~ x1 + x2 + x1:x2"
get_variables(form)
form <- "y ~ x1 + x2 + x1*x2"
get_variables(form)}
}
