% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_auxiliary.r
\name{mnom_liab2cat}
\alias{mnom_liab2cat}
\title{Multinomial liability to category}
\usage{
mnom_liab2cat(liability, categories)
}
\arguments{
\item{liability}{Matrix of liabilities (n_cases x K-1 for K categories)}

\item{categories}{Vector of category labels}
}
\value{
Character vector of sampled categories
}
\description{
Converts multivariate liability scale to categorical outcomes
}
