% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_auxiliary.r
\name{ordinal_liab2cat}
\alias{ordinal_liab2cat}
\title{Ordinal liability to category}
\usage{
ordinal_liab2cat(liability, n_cats, threshold_spread = 1.5)
}
\arguments{
\item{liability}{Vector of latent liabilities (n_cases)}

\item{n_cats}{Number of ordered categories}

\item{threshold_spread}{Controls spread of thresholds (default 1.5)}
}
\value{
Integer vector of ordered categories (1, 2, ..., n_cats)
}
\description{
Converts latent liability to ordered categorical outcomes (1, 2, 3, ...)
using cumulative probit/logit model with evenly spaced thresholds
}
