% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_plots.R
\name{plot_acf_convergence}
\alias{plot_acf_convergence}
\title{plot_acf_convergence}
\usage{
plot_acf_convergence(conv_object, variables = NULL, ...)
}
\arguments{
\item{conv_object}{Object of class bace_convergence}

\item{variables}{Variables to plot}

\item{...}{Additional plotting arguments}
}
\value{
Invisibly returns \code{NULL}; called for its side effect of drawing
autocorrelation (ACF) panels (one per variable) on the current graphics device.
}
\description{
Plot autocorrelation diagnostics
}
