% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_plots.R
\name{plot_energy_convergence}
\alias{plot_energy_convergence}
\title{plot_energy_convergence}
\usage{
plot_energy_convergence(conv_object, ...)
}
\arguments{
\item{conv_object}{Object of class bace_convergence}

\item{...}{Additional plotting arguments}
}
\value{
Invisibly returns \code{NULL}; called for its side effect of drawing
the energy-distance trace plot on the current graphics device.
}
\description{
Plot energy distance evolution
}
