% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_plots.R
\name{plot_wasserstein_convergence}
\alias{plot_wasserstein_convergence}
\title{plot_wasserstein_convergence}
\usage{
plot_wasserstein_convergence(conv_object, variables = NULL, ...)
}
\arguments{
\item{conv_object}{Object of class bace_convergence}

\item{variables}{Variables to plot}

\item{...}{Additional plotting arguments}
}
\value{
Invisibly returns \code{NULL}; called for its side effect of drawing
Wasserstein-distance panels (one per variable) on the current graphics device.
}
\description{
Plot Wasserstein distance evolution for each variable
}
