% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{predict.bace}
\alias{predict.bace}
\title{predict.bace}
\usage{
\method{predict}{bace}(model, type = NULL, ...)
}
\arguments{
\item{model}{A MCMCglmm model object}
}
\value{
A list of priors for the MCMCglmm model.
}
\description{
Function creates a predcition from MCMCglmm model
}
