% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bace_final_imp.R
\name{print.bace_final}
\alias{print.bace_final}
\title{print.bace_final}
\usage{
\method{print}{bace_final}(x, ...)
}
\arguments{
\item{x}{Object of class bace_final}

\item{...}{Additional arguments}
}
\value{
Returns \code{x} invisibly; called for its side effect of printing
a summary of the \code{bace_final} object to the console.
}
\description{
Print method for bace_final objects
}
