% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_simBACE.R
\name{print_sim_bace_summary}
\alias{print_sim_bace_summary}
\title{Print summary of sim_bace output}
\usage{
print_sim_bace_summary(sim_output)
}
\arguments{
\item{sim_output}{Output from sim_bace function}
}
\value{
The input \code{sim_output} list, invisibly. Called for its side
effect of printing a summary of the simulation to the console.
}
\description{
Prints a summary of the simulated data
}
