% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_simBACE.R
\name{print_simBACE_summary}
\alias{print_simBACE_summary}
\title{Print summary of simBACE output}
\usage{
print_simBACE_summary(sim_output)
}
\arguments{
\item{sim_output}{Output from simBACE function}
}
\description{
Prints a summary of the simulated data
}
