% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_auxiliary.r
\name{sample_random_effects}
\alias{sample_random_effects}
\title{Sample random effects}
\usage{
sample_random_effects(sigma2, n, cor_matrix = NULL)
}
\arguments{
\item{sigma2}{Variance of the random effect}

\item{n}{Number of levels}

\item{cor_matrix}{Correlation matrix (for phylogenetic effects)}
}
\value{
Vector of random effect values
}
\description{
Generates random effect coefficients for a given variance
}
