% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_simBACE.R
\name{sim_bace_gaussian}
\alias{sim_bace_gaussian}
\title{Quick Gaussian simulation}
\usage{
sim_bace_gaussian(
  n_predictors = 3,
  n_cases = 200,
  n_species = 75,
  phylo_signal = 0,
  beta_sparsity = 0.7
)
}
\arguments{
\item{n_predictors}{Number of gaussian predictors}

\item{n_cases}{Number of observations}

\item{n_species}{Number of species}

\item{phylo_signal}{Single phylogenetic signal value applied to all variables}

\item{beta_sparsity}{Proportion of coefficients in beta_matrix to set to zero (default 0.7)}
}
\value{
simBACE output list
}
\description{
Simplified wrapper for gaussian response simulation
}
