% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_simBACE.R
\name{sim_bace_poisson}
\alias{sim_bace_poisson}
\title{Quick Poisson simulation}
\usage{
sim_bace_poisson(
  n_predictors = 3,
  n_cases = 200,
  n_species = 75,
  phylo_signal = 0
)
}
\arguments{
\item{n_predictors}{Number of gaussian predictors}

\item{n_cases}{Number of observations}

\item{n_species}{Number of species}

\item{phylo_signal}{Single phylogenetic signal value applied to all variables}
}
\value{
simBACE output list
}
\description{
Simplified wrapper for count response simulation
}
