% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_auxiliary.r
\name{sim_tree}
\alias{sim_tree}
\title{Simulate phylogenetic tree and assign cases to species}
\usage{
sim_tree(n_species, birth, death, n_cases, str_len = 5)
}
\arguments{
\item{n_species}{Number of species on the tree}

\item{birth}{Birth rate for tree simulation}

\item{death}{Death rate for tree simulation}

\item{n_cases}{Number of observations to generate}

\item{str_len}{Length of random species name strings}
}
\value{
List containing tree and case-to-species assignment
}
\description{
Simulates a phylogenetic tree using birth-death process
and assigns observations to species
}
