% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_auxiliary.r
\name{var_name_gen}
\alias{var_name_gen}
\title{Generate default variable names}
\usage{
var_name_gen(n_predictors)
}
\arguments{
\item{n_predictors}{Number of predictor variables}
}
\value{
Character vector of variable names (y, x1, x2, ...)
}
\description{
Generates default variable names if not provided
}
