% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_functions.R
\docType{package}
\name{BACE-package}
\alias{BACE}
\alias{BACE-package}
\title{BACE: Bayesian Augmentation using Chained Equations (BACE)}
\description{
Bayesian Augmentation using Chained Equations (BACE) for imputing missing data in comparative datasets.
}
\author{
\strong{Maintainer}: Shinichi Nakagawa \email{s.nakagawa@unsw.edu.au}

Authors:
\itemize{
  \item Szymek Drobniak \email{szymek.drobniak@gmail.com}
  \item Daniel Noble \email{daniel.noble@anu.edu.au}
}

}
\keyword{internal}
