% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bace_final_imp.R
\name{bace_final_imp}
\alias{bace_final_imp}
\title{bace_final_imp}
\usage{
bace_final_imp(
  bace_object,
  fixformula,
  ran_phylo_form,
  phylo,
  nitt = 6000,
  thin = 5,
  burnin = 1000,
  n_final = 10,
  species = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{bace_object}{An object of class 'bace' from bace_imp}

\item{fixformula}{A character string or list specifying the fixed effects formulas}

\item{ran_phylo_form}{A character string specifying the random effects and phylogenetic structure formula}

\item{phylo}{A phylogenetic tree of class 'phylo' from the ape package}

\item{nitt}{An integer or list specifying the number of MCMC iterations. Default is 6000}

\item{thin}{An integer or list specifying the thinning rate. Default is 5}

\item{burnin}{An integer or list specifying the burn-in iterations. Default is 1000}

\item{n_final}{An integer specifying the number of final imputation runs. Default is 10}

\item{species}{A logical indicating whether to decompose phylogenetic and non-phylogenetic species effects. Default is FALSE}

\item{verbose}{A logical indicating whether to print progress messages. Default is TRUE}

\item{...}{Additional arguments passed to modeling functions}
}
\value{
A list of class 'bace_final' containing:
\itemize{
\item all_models: List of length n_final, each containing models for all variables
\item all_datasets: List of length n_final with imputed datasets
\item formulas: The formulas used
\item types: Variable types
\item phylo_ran: Phylogenetic random effects structure
\item call: The function call
}
}
\description{
Run final imputation iterations starting from converged data, saving all models
}
\examples{
\dontrun{
# After running bace_imp and checking convergence
result <- bace_imp(fixformula = "y ~ x1 + x2", ...)
final <- bace_final_imp(result, fixformula = "y ~ x1 + x2", ...)
}
}
