% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bace_imp.R
\name{bace_imp}
\alias{bace_imp}
\title{bace_imp}
\usage{
bace_imp(
  fixformula,
  ran_phylo_form,
  phylo,
  data,
  runs = 10L,
  nitt = 50000,
  thin = 10,
  burnin = 1000,
  ...
)
}
\arguments{
\item{fixformula}{A character string specifying the fixed effects formula that is of major interest or a list of formulas that one wishes to estimate. This should be of the form: y ~ x. Providing a list gives users complex flexibility in the types of models fit for the different variables. Note that users can also include interaction terms using the standard R formula syntax (e.g., x1*x2).}

\item{ran_phylo_form}{A character string specifying the random effects and phylogenetic structure formula used in the model.}

\item{phylo}{A phylogenetic tree of class 'phylo' from the ape package.}

\item{data}{A data frame containing the dataset with missing values to be imputed.}

\item{runs}{An integer specifying the number of imputation iterations to perform. Default is 10.}

\item{...}{Additional arguments to be passed to the underlying modeling functions.}

\item{species}{A logical indicating whether species is included as a random effect.}
}
\value{
A list containing imputed datasets and model summaries.
}
\description{
Function to perform Bayesian imputation using BACE for missing data in a dataset
}
\examples{
\dontrun{
phylo <- force.ultrametric(ape::rtree(5)) # Example phylogenetic tree with 5 tips
data <- data.frame(y = c(1,2,3,NA,5), x1 = factor(c("A","B","A","B","A")), x2 = c(10,20,30,40,50), Species = phylo$tip.label)
bace_imp(fixformula = "y ~ x1 + x2", ran_phylo_form = "~ 1 |Species", phylo = phylo, data = data)
bace_imp(fixformula = list("y ~ x1 + x2", "x2 ~ x1"), ran_phylo_form = "~ 1 |Species", phylo = phylo, data = data, runs = 20)
}
}
