% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bace_imp.R
\name{bace_imp}
\alias{bace_imp}
\title{bace_imp}
\usage{
bace_imp(
  fixformula,
  ran_phylo_form,
  phylo,
  data,
  nitt = 6000,
  thin = 5,
  burnin = 1000,
  runs = 10,
  ...
)
}
\arguments{
\item{fixformula}{A character string specifying the fixed effects formula that is of major interest or a list of formulas that one wishes to estimate. This should be of the form: y ~ x. Providing a list gives users complex flexibility in the types of models fit for the different variables. Note that users can also include interaction terms using the standard R formula syntax (e.g., x1*x2).}

\item{ran_phylo_form}{A character string specifying the random effects and phylogenetic structure formula used in the model.}

\item{phylo}{A phylogenetic tree of class 'phylo' from the ape package.}

\item{data}{A data frame containing the dataset with missing values to be imputed.}

\item{nitt}{An integer specifying the number of iterations to run the MCMC algorithm. Default is 6000.}

\item{thin}{An integer specifying the thinning rate for the MCMC algorithm. Default is 5.}

\item{burnin}{An integer specifying the number of iterations to discard as burnin. Default is 1000.}

\item{runs}{An integer specifying the number of imputation iterations to perform. Default is 10.}

\item{...}{Additional arguments to be passed to the underlying modeling functions.}
}
\value{
A list containing imputed datasets and model summaries.
}
\description{
Function to perform Bayesian imputation using BACE for missing data in a dataset
}
\examples{
\dontrun{
set.seed(123)
# Example phylogenetic tree with 30 tips
phylo <- phytools::force.ultrametric(ape::rtree(30))
phylo  <- ape::compute.brlen(phylo, method = "Grafen")
data <- data.frame(
  y = rpois(30, lambda = 5),
  x1 = factor(rep(c("A", "B","A"), length.out = 30)),
  x2 = rnorm(30, 10, 2),
  x3 = factor(rep(c("A", "B", "C", "D", "E"), length.out = 30)),
  x4 = factor(
    rep(c("A", "B", "C", "D", "E"), length.out = 30),
    levels = c("B", "A", "C", "D", "E"),
    ordered = TRUE
  ),
  Species = phylo$tip.label
)
# Introduce some missing data
missing_indices <- sample(1:30, 10)
data$y[missing_indices] <- NA
data$x1[sample(1:30, 5)] <- NA
data$x2[sample(1:30, 5)] <- NA	
data$x3[sample(1:30, 5)] <- NA
data$x4[sample(1:30, 5)] <- NA	
# Run BACE imputation
bace_imp(
  fixformula = "y ~ x1 + x2",
  ran_phylo_form = "~ 1 |Species",
  phylo = phylo,
  data = data
)
bace_imp(
  fixformula = list(
    "y ~ x1 + x2", "x2 ~ x1", "x1 ~ x2",
    "x3 ~ x1 + x2", "x4 ~ x1 + x2"
  ),
  ran_phylo_form = "~ 1 |Species",
  phylo = phylo,
  data = data,
  runs = 5
)
}
}
