% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bace_option_setter_getter.r
\name{bace_options}
\alias{bace_options}
\title{Get or set BACE options}
\usage{
bace_options(..., .reset = FALSE)
}
\arguments{
\item{...}{Named options to set. If empty, returns current options.}

\item{.reset}{Logical; if TRUE, reset all mypkg.* options (in this session) to defaults.}
}
\value{
A named list of effective option values (after defaults + overrides).
}
\description{
Get or set BACE options
}
\examples{
\dontrun{
# get current options
bace_options()
#' # set verbose to TRUE
bace_options(verbose = TRUE)
bace_options()
# reset all options to defaults
bace_options(.reset = TRUE)
# user overriden option
getOption("BACE.verbose")
}
}
