% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_functions.R
\name{check_type}
\alias{check_type}
\title{check_type}
\usage{
check_type(x)
}
\arguments{
\item{x}{A vector of data}
}
\value{
Returns a character string specifying the class of the variable.
}
\description{
Function takes a variable in a dataframe and checks / classifies what type of variable it is (binary, continuous, count, categorical, ordered_categorical) so that it can be modeled appropriately.
}
\examples{
\dontrun{
x1 <- as.factor(c(1,0,1,1,0,NA))
check_type(x1) # returns "binary"
x1.1 <- c(1,0,1,1,0,NA)
check_type(x1.1) # returns "warning" about binary numeric coding
x2 <- c(1.5, 2.3, 3.1, NA, 4.0)
check_type(x2) # returns "continuous"
x3 <- c(0,1,2,3,4,NA)	
check_type(x3) # returns "count"
x4 <- as.factor(c("A","B","C","A","B",NA))
check_type(x4) # returns "categorical"
x5 <- as.ordered(c("low","medium","high","medium",NA))
check_type(x5) # returns "ordered_categorical"
}
}
