% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{.build_formula_string_random}
\alias{.build_formula_string_random}
\title{.build_formula_string_random}
\usage{
.build_formula_string_random(ran_phylo_form)
}
\arguments{
\item{ran_phylo_form}{A character string specifying the random effects and phylogenetic structure formula used in the model.}
}
\value{
A formula	 object
}
\description{
Function takes a string specifying a random effects formula and converts this to a formula object to be used in the models.
}
\examples{
\dontrun{
.build_formula_string_random("~ 1 | Species")
}
}
