% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{.pred_count}
\alias{.pred_count}
\title{.pred_count}
\usage{
.pred_count(model)
}
\arguments{
\item{model}{A MCMCglmm model object}
}
\value{
A data frame with two columns: post_mean and post_sd (row per observation)
}
\description{
Posterior mean and posterior SD of fitted values for a Poisson (log-link) MCMCglmm model
}
