% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{.predict_bace}
\alias{.predict_bace}
\title{.predict_bace}
\usage{
.predict_bace(model, dat_prep, response_var, type = NULL, sample = FALSE, ...)
}
\arguments{
\item{model}{A MCMCglmm model object}

\item{dat_prep}{A list containing the prepared data frame and attributes for continuous variables.}

\item{response_var}{A string specifying the name of the response variable.}

\item{type}{A string that specifies the type of model to fit.}

\item{sample}{A logical indicating whether to sample from the distribution for categorical/ordinal variables.}

\item{...}{Additional arguments (not used).}
}
\value{
A vector of predicted values from the MCMCglmm model.
}
\description{
Function creates a predcition from MCMCglmm model
}
