% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_functions.R
\name{.summarise_var_types}
\alias{.summarise_var_types}
\title{summarise_var_types}
\usage{
.summarise_var_types(df, store_levels = TRUE, max_levels_store = 200)
}
\arguments{
\item{df}{A data frame containing the dataset.}

\item{store_levels}{A logical indicating whether to store levels for factor and character variables. Default is TRUE.}

\item{max_levels_store}{An integer specifying the maximum number of levels to store. Default is 200.}
}
\value{
A data frame summarizing the types and characteristics of each variable, the number of unique levels and whether it has missing data.
}
\description{
Function summarizes the types and characteristics of variables in a dataframe.
}
\examples{
\dontrun{
data <- data.frame(y = c(1,2,3,NA,5), x1 = factor(c("A","B","A","B","A")), x2 = c(10,20,30,NA,50))
summarise_var_types(data)
}
}
