% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_functions.R
\name{get_type}
\alias{get_type}
\title{get_type}
\usage{
get_type(data)
}
\arguments{
\item{fixed_form}{The fixed effect formula}

\item{cluster_name}{The name of the cluster variable}

\item{random_form}{The random effect formula}
}
\value{
A list of variable names and their correpsonding class.
}
\description{
Function takes a dataframe and identifies the class of variables (type)
}
\examples{
{
data(iris)}
}
