% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_functions.R
\name{get_variables}
\alias{get_variables}
\title{get_variables}
\usage{
get_variables(x, fix = TRUE)
}
\arguments{
\item{x}{A string}

\item{fix}{A logical value. If TRUE, the function returns the fixed effects variables. If FALSE, the function returns the random effects variables.}
}
\value{
A vector of variable names. This vector is used to subset out the 'fixed effects' and 'random effect' columns within a given dataframe.
}
\description{
Function takes a formula string and identifies the variables in the formula that relate to the data
}
\examples{
{
# All should return the same three variables: y, x1, x2
form <- "y ~ x1 + x2"
get_variables(form)
form <- "y ~ x1 * x2"
get_variables(form)
form <- "y ~ x1 + x2 + x1 : x2"
get_variables(form)
form <- "y ~ x1 + x2 + x1:x2"
get_variables(form)
form <- "y ~ x1 + x2 + x1*x2"
get_variables(form)
form <- "~ 1 + X1|Species"
get_variables(form, fix = FALSE)
form <- "~ 1 |Species"
get_variables(form, fix = FALSE)}
}
