% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_plots.R
\name{plot_convergence_summary}
\alias{plot_convergence_summary}
\title{plot_convergence_summary}
\usage{
plot_convergence_summary(conv_object, ...)
}
\arguments{
\item{conv_object}{Object of class bace_convergence}

\item{...}{Additional plotting arguments}
}
\description{
Create a comprehensive summary plot of convergence diagnostics
}
