% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bace.R
\name{print.bace_complete}
\alias{print.bace_complete}
\title{print.bace_complete}
\usage{
\method{print}{bace_complete}(x, ...)
}
\arguments{
\item{x}{Object of class bace_complete}

\item{...}{Additional arguments}
}
\value{
Returns \code{x} invisibly; called for its side effect of printing
a summary of the \code{bace_complete} object to the console.
}
\description{
Print method for bace_complete objects
}
