% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_diagnostics.R
\name{print.bace_convergence}
\alias{print.bace_convergence}
\title{print.bace_convergence}
\usage{
\method{print}{bace_convergence}(x, ...)
}
\arguments{
\item{x}{Object of class bace_convergence}

\item{...}{Additional arguments}
}
\description{
Print method for bace_convergence objects
}
