% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{print.bace_diagnostics}
\alias{print.bace_diagnostics}
\title{print.bace_diagnostics}
\usage{
\method{print}{bace_diagnostics}(x, ...)
}
\arguments{
\item{x}{A bace_diagnostics object from .check_mcmc_diagnostics()}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for MCMC diagnostics from bace_imp models
}
