% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_posteriors.R
\name{print.bace_pooled}
\alias{print.bace_pooled}
\title{print.bace_pooled}
\usage{
\method{print}{bace_pooled}(x, ...)
}
\arguments{
\item{x}{Object of class bace_pooled}

\item{...}{Additional arguments}
}
\description{
Print method for bace_pooled objects
}
