% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_posteriors.R
\name{print.bace_pooled_MCMCglmm}
\alias{print.bace_pooled_MCMCglmm}
\title{print.bace_pooled_MCMCglmm}
\usage{
\method{print}{bace_pooled_MCMCglmm}(x, ...)
}
\arguments{
\item{x}{Object of class bace_pooled_MCMCglmm}

\item{...}{Additional arguments passed to MCMCglmm print method}
}
\description{
Print method for individual pooled MCMCglmm model objects
}
