% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_posteriors.R
\name{summary.bace_pooled_MCMCglmm}
\alias{summary.bace_pooled_MCMCglmm}
\title{summary.bace_pooled_MCMCglmm}
\usage{
\method{summary}{bace_pooled_MCMCglmm}(object, ...)
}
\arguments{
\item{object}{Object of class bace_pooled_MCMCglmm}

\item{...}{Additional arguments passed to MCMCglmm summary method}
}
\description{
Summary method for individual pooled MCMCglmm model objects
}
